FIXME test: external add dir
